Refactor sort component
